Controller Etat (test ok)

+++ src/main/java/slam/itis/NoteDeFrais/service/EtatService.java
@@ -3,6 +3,7 @@ package slam.itis.NoteDeFrais.service;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import slam.itis.NoteDeFrais.Repository.EtatRepository;
@@ -13,6 +14,7 @@ public class EtatService {
 
     private final EtatRepository etatRepository;
 
+    @Autowired
     public EtatService(EtatRepository etatRepository) {
         this.etatRepository = etatRepository;
     }
@@ -30,6 +32,13 @@ public class EtatService {
     }
 
     public void deleteEtatById(Long id) {
+        if (!etatRepository.existsById(id)) {
+            throw new IllegalArgumentException("État avec ID " + id + " introuvable.");
+        }
         etatRepository.deleteById(id);
+    }
+
+    public Etat getEtatByLibelle(String libelle) {
+        return etatRepository.findByLibelle(libelle);
     }
 }
